requirements.txt
requests
fastapi
uvicorn[standard]
pymongo
python-dotenv
email-validator
jinja2
python-multipart
python-jose
passlib[bcrypt]
openai
celery
redis
httpx
